html/editing/dnd/drop/events-textarea-manual.tentative.html

<!doctype html>
<meta charset=utf-8>
<title>Selection drag and drop: events for &lt;textarea></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div><input id=a data-select="0,7" value="Drag me" data-expected-events="beforeinput,input"></div>
<div><textarea id=b placeholder="...to here" data-expected-events="drop,beforeinput,textInput,input"></textarea></div>
<script src="/uievents/textInput/support/common.js"></script>
<script src="support/events.js"></script>
